requirements.txt
# A2A Protocol SDK
a2a-sdk[http-server]>=0.3.0

# AI for natural language processing and analysis
google-genai>=1.0.0

# HTTP client for Dynatrace API calls
httpx>=0.27.0

# Environment variables
python-dotenv>=1.0.0

# Server
uvicorn>=0.30.0
